# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.loginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/model.AdminUser'
    type: object
  api.updateArticleRequest:
    properties:
      content_md:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  api.updateProfileRequest:
    properties:
      email:
        type: string
    type: object
  model.AdminUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  model.Article:
    properties:
      content_md:
        type: string
      cover_asset_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      like_count:
        type: integer
      published_at:
        type: string
      repo_analysis_id:
        type: integer
      slug:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      word_count:
        type: integer
    type: object
...
  resp.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  service.ArticleDetail:
    properties:
      content_md:
        type: string
      cover_asset_id:
        type: integer
      created_at:
        type: string
      figures:
        items:
          $ref: '#/definitions/service.FigureMeta'
        type: array
      id:
        type: integer
      like_count:
        type: integer
      published_at:
        type: string
      repo_analysis_id:
        type: integer
      slug:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      word_count:
        type: integer
    type: object
  service.FigureMeta:
    properties:
      article_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      kind:
        type: string
      placeholder:
        type: string
      title:
        type: string
    type: object
  service.PageData:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  service.PortalArticleDetail:
    properties:
      content_md:
        type: string
      cover:
        $ref: '#/definitions/service.PortalCover'
      cover_asset_id:
        type: integer
      created_at:
        type: string
      figures:
        items:
          $ref: '#/definitions/service.PortalFigure'
        type: array
      id:
        type: integer
      like_count:
        type: integer
      published_at:
        type: string
      repo_analysis_id:
        type: integer
      slug:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      word_count:
        type: integer
    type: object
  service.PortalCover:
    properties:
      id:
        type: integer
      kind:
        type: string
      title:
        type: string
    type: object
  service.PortalFigure:
    properties:
      id:
        type: integer
      kind:
        type: string
      placeholder:
        type: string
      title:
        type: string
    type: object
  service.PortalLikeResult:
    properties:
      duplicated:
        type: boolean
      like_count:
        type: integer
    type: object
  service.PortalRelatedArticle:
    properties:
      published_at:
        type: string
      slug:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    AI 博客生成平台后端：管理端（任务/文章/RAG 向量索引）+ 前台门户（文章浏览/RAG 问答）。
    统一响应结构 resp.Envelope{code, message, data}，code=0 表示成功；业务错误 code 为 HTTP 状态码。
    /metrics 为 Prometheus 抓取端点，不在 /api/v1 之下，本文档未收录。
  title: AI 博客平台 API
  version: "1.0"
paths:
  /articles:
    get:
      parameters:
      - description: 状态过滤：draft / published，可省略
        in: query
        name: status
        type: string
      - description: 页码，默认 1
        in: query
        name: page
        type: integer
      - description: 每页条数，默认 10，上限 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/service.PageData'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.Article'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 文章分页列表（管理端）
      tags:
      - 文章
  /articles/{id}:
    get:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/service.ArticleDetail'
              type: object
      security:
      - BearerAuth: []
      summary: 文章详情
      tags:
      - 文章
    put:
      consumes:
      - application/json
      description: 已发布文章保存后自动退回 draft，需重新发版。
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标题/摘要/正文/标签
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.Article'
              type: object
      security:
      - BearerAuth: []
      summary: 编辑文章
      tags:
      - 文章
  /articles/{id}/offline:
    post:
      description: published → draft；同时清理 Redis 前台缓存并删除该篇 RAG 向量。
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.Article'
              type: object
      security:
      - BearerAuth: []
      summary: 下线文章
      tags:
      - 文章
  /articles/{id}/publish:
    post:
      description: draft → published；同时触发 Redis 前台缓存清理与 RAG 向量化（异步）。
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.Article'
              type: object
      security:
      - BearerAuth: []
      summary: 发版文章
      tags:
      - 文章
  /articles/{id}/regenerate-figures:
    post:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/service.ArticleDetail'
              type: object
      security:
      - BearerAuth: []
      summary: 重新生成配图
      tags:
      - 文章
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 账号密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.loginResponse'
              type: object
        "401":
          description: 账号或密码错误
          schema:
            $ref: '#/definitions/resp.Envelope'
      summary: 管理端登录
      tags:
      - 认证
  /auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.AdminUser'
              type: object
      security:
      - BearerAuth: []
      summary: 查看当前管理员信息
      tags:
      - 认证
    put:
      consumes:
      - application/json
      parameters:
      - description: 邮箱
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.AdminUser'
              type: object
      security:
      - BearerAuth: []
      summary: 更新当前管理员邮箱
      tags:
      - 认证
  /portal/articles:
    get:
      parameters:
      - description: 页码，默认 1
        in: query
        name: page
        type: integer
      - description: 每页条数，默认 10，上限 100
        in: query
        name: page_size
        type: integer
      - description: 按标签过滤
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/service.PageData'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.Article'
                        type: array
                    type: object
              type: object
      summary: 文章分页列表（前台）
      tags:
      - 门户
  /portal/articles/{slug}:
    get:
      parameters:
      - description: 文章 slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/service.PortalArticleDetail'
              type: object
      summary: 文章详情（前台）
      tags:
      - 门户
  /portal/articles/{slug}/like:
    post:
      parameters:
      - description: 文章 slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/service.PortalLikeResult'
              type: object
      summary: 文章点赞
      tags:
      - 门户
  /portal/articles/{slug}/related:
    get:
      parameters:
      - description: 文章 slug
        in: path
        name: slug
        required: true
        type: string
      - description: 返回条数，默认 4
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  properties:
                    items:
                      items:
                        $ref: '#/definitions/service.PortalRelatedArticle'
                      type: array
                  type: object
              type: object
      summary: 相关文章推荐
      tags:
      - 门户
...
  /portal/figures/{file}:
    get:
      description: 前台 Markdown 内以 /api/v1/portal/figures/xx.svg 引用；非 JSON，直接输出 SVG。
      parameters:
      - description: 配图文件名，如 12.svg 或 12
        in: path
        name: file
        required: true
        type: string
      produces:
      - image/svg+xml
      responses:
        "200":
          description: image/svg+xml
          schema:
            type: string
        "400":
          description: 非法配图 id
          schema:
            $ref: '#/definitions/resp.Envelope'
      summary: SVG 配图
      tags:
      - 门户
...
securityDefinitions:
  BearerAuth:
    description: '管理端接口需在请求头携带登录接口返回的 JWT：`Authorization: Bearer <token>`。'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"